oBG2 troll handling not needed on EEs
Troll already has trollreg.itm, which handles the detah-revive stuff on EEs. As such I removed scripting and dead troll cre
file when installed on EEs as it's possible, though unlikely, that they may interfere with EE's method.

Also added the gland to Roger's not-dead sea troll, as rogtro02 is unused on EEs.